Redo the main menu
This commit improves the main menu display.

- The logic for displaying main menu buttons was rewritten using Godot
mechanisms. Now buttons are displayed using a UI container, and an
anchor is used to position the button container inside the scene.

- Title screen art preserves its ratio when resizing the screen.

- The [Tool] annotation usage was fixed. Previously, the MainMenu
class was marked with the tool annotation. This led to issues when
editing a scene, since MainMenu requires access to the Godot singleton
which is inaccessible within the editor. It also led to the problem
where some of the MainMenu export properties were reset on opening a
scene: this was caused by the fact that non-tool classes can't be set
as export properties of a tool. Now only MenuButtonContainer is marked
as [Tool], so the buttons are correctly displayed in the scene editor.

- New title screen art has been added! To prepare for the upcoming
Dutch preview, I decided to choose new art for the menu. It is a
painting called "The Great Market in Haarlem" by Gerrit
Berckheyde (1696). It doesn't have too many details and conveys some
of the themes of the upcoming release (population, buildings,
culture...), so I think it works pretty well.
