--- C7/MainMenu.cs
@@ -1,169 +1,0 @@
-using Godot;
-using System;
-using C7Engine;
-using C7GameData;
-using C7GameData.Save;
-using Serilog;
-
-[Tool]
-public partial class MainMenu : Node2D {
-	private ILogger log;
-
-	readonly int BUTTON_LABEL_OFFSET = 0;
-
-	ImageTexture InactiveButton;
-	ImageTexture HoverButton;
-	TextureRect MainMenuBackground;
-	[Export]
-	Civ3FileDialog LoadDialog;
-	[Export]
-	Button SetCiv3Home;
-	[Export]
-	FileDialog SetCiv3HomeDialog;
-	[Export]
-	Civ3FileDialog LoadScenarioDialog;
-	GlobalSingleton Global;
-
-	Button toggleGraphicsButton;
-
-	readonly int MENU_OFFSET_FROM_TOP = 180;
-	readonly int MENU_OFFSET_FROM_LEFT = 180;
-
-	// Called when the node enters the scene tree for the first time.
-	public override void _Ready() {
-		log = LogManager.ForContext<MainMenu>();
-		log.Debug("enter MainMenu._Ready");
-
-		DisplayServer.WindowSetTitle("C7 - Godot 4");
-
-		DisplayTitleScreen();
-	}
-
-	private void DisplayTitleScreen() {
-		try {
-			SetMainMenuBackground();
-
-			// To pass data between scenes, putting path string in a global singleton and reading it later in createGame
-			Global = GetNode<GlobalSingleton>("/root/GlobalSingleton");
-			Global.ResetLoadGamePath();
-
-			LoadDialog.SetDirectoryForLoading(@"Conquests/Saves");
-			LoadScenarioDialog.SetDirectoryForLoading(@"Conquests/Scenarios");
-
-			InactiveButton = TextureLoader.Load("ui.button.inactive");
-			HoverButton = TextureLoader.Load("ui.button.hover");
-
-			AddButton("New Game", 0, GoToWorldSetup);
-			AddButton("Quick Start", 35, GoToWorldSetup);
-			AddButton("Tutorial", 70, StartGame);
-			AddButton("Load Game", 105, LoadGame);
-			AddButton("Load Scenario", 140, LoadScenario);
-			AddButton("Hall of Fame", 175, HallOfFame);
-			(_, toggleGraphicsButton) = AddButton("Turn on C7 Graphics", 210, () => {
-				if (toggleGraphicsButton.Text == "Turn on C7 Graphics") {
-					toggleGraphicsButton.Text = "Turn on Civ3 Graphics";
-				} else {
-					toggleGraphicsButton.Text = "Turn on C7 Graphics";
-				}
-				TextureLoader.ToggleModernGraphics();
-			});
-			AddButton("Preferences", 245, Preferences);
-			AddButton("Audio Preferences", 280, Preferences);
-			AddButton("Credits", 315, showCredits);
-			AddButton("Exit", 350, _on_Exit_pressed);
-
-			// Hide select home folder if valid path is present as proven by reaching this point in code
-			SetCiv3Home.Visible = false;
-		} catch (Exception ex) {
-			log.Error(ex, "Could not set up the main menu");
-			GetNode<Label>("CanvasLayer/Label").Visible = true;
-			GetNode<ColorRect>("CanvasLayer/ColorRect").Visible = true;
-		}
-	}
-
-	private void SetMainMenuBackground() {
-		ImageTexture TitleScreenTexture = Util.LoadTextureFromC7JPG("Art/Title_Screen.jpg");
-		MainMenuBackground = GetNode<TextureRect>("CanvasLayer/MainMenuBackground");
-		MainMenuBackground.StretchMode = TextureRect.StretchModeEnum.Scale;
-		MainMenuBackground.Texture = TitleScreenTexture;
-	}
-
-	private (TextureButton, Button) AddButton(string label, int verticalPosition, Action action) {
-		TextureButton newButton = new TextureButton();
-		newButton.TextureNormal = InactiveButton;
-		newButton.TextureHover = HoverButton;
-		newButton.SetPosition(new Vector2(MENU_OFFSET_FROM_LEFT, MENU_OFFSET_FROM_TOP + verticalPosition));
-		MainMenuBackground.AddChild(newButton);
-		newButton.Pressed += action;
-
-		Theme theme = new Theme();
-		theme.SetFontSize("font_size", "Button", 14);
-		Button newButtonLabel = new Button();
-		newButtonLabel.Theme = theme;
-		newButtonLabel.Text = label;
-
-		newButtonLabel.SetPosition(new Vector2(MENU_OFFSET_FROM_LEFT + 25, MENU_OFFSET_FROM_TOP + verticalPosition + BUTTON_LABEL_OFFSET));
-		MainMenuBackground.AddChild(newButtonLabel);
-		newButtonLabel.Pressed += action;
-		return (newButton, newButtonLabel);
-	}
-
-	public void GoToWorldSetup() {
-		PlayButtonPressedSound();
-		GetTree().ChangeSceneToFile("res://UIElements/NewGame/world_setup.tscn");
-	}
-
-	public void StartGame() {
-		log.Information("start game button pressed");
-		PlayButtonPressedSound();
-		GetTree().ChangeSceneToFile("res://C7Game.tscn");
-	}
-
-	public void LoadGame() {
-		log.Information("load game button pressed");
-		PlayButtonPressedSound();
-		LoadDialog.Popup();
-	}
-
-	public void LoadScenario() {
-		log.Information("load scenario button pressed");
-		PlayButtonPressedSound();
-		LoadScenarioDialog.Popup();
-	}
-
-	public void showCredits() {
-		log.Information("credits button pressed");
-		GetTree().ChangeSceneToFile("res://Credits.tscn");
-	}
-
-	public void HallOfFame() {
-		PlayButtonPressedSound();
-	}
-
-	public void Preferences() {
-		PlayButtonPressedSound();
-	}
-
-	public void _on_Exit_pressed() {
-		GetTree().Quit(); // no need to notify the scene tree
-	}
-
-	private void PlayButtonPressedSound() {
-		AudioStreamWav wav = Util.LoadWAVFromDisk(Util.Civ3MediaPath("Sounds/Button1.wav"));
-		AudioStreamPlayer player = GetNode<AudioStreamPlayer>("CanvasLayer/SoundEffectPlayer");
-		player.Stream = wav;
-		player.Play();
-	}
-
-	private void _on_SetCiv3Home_pressed() {
-		SetCiv3HomeDialog.Popup();
-	}
-
-	private void _on_SetCiv3HomeDialog_dir_selected(string path) {
-		Util.Civ3Root = path;
-		C7Settings.SetValue("locations", "civ3InstallDir", path);
-		C7Settings.SaveSettings();
-		// This function should only be reachable if DisplayTitleScreen failed on previous runs, so should be OK to run here
-		DisplayTitleScreen();
-	}
-}
