--- C7/MainMenu.tscn
@@ -1,111 +1,0 @@
-[gd_scene load_steps=9 format=3 uid="uid://gnt0y3og7nk6"]
-
-[ext_resource type="Script" path="res://MainMenu.cs" id="1"]
-[ext_resource type="Script" path="res://MainMenuMusicPlayer.cs" id="3"]
-[ext_resource type="Theme" uid="uid://b0jisy3avwkxf" path="res://MainMenuTheme.tres" id="4"]
-[ext_resource type="PackedScene" uid="uid://bvncm1bgcii5e" path="res://UIElements/civ3_file_dialog.tscn" id="4_x23yr"]
-[ext_resource type="Script" path="res://UIElements/Civ3TextureRect.cs" id="5"]
-
-[sub_resource type="StyleBoxFlat" id="StyleBoxFlat_gmerc"]
-
-[sub_resource type="StyleBoxFlat" id="StyleBoxFlat_vbij8"]
-
-[sub_resource type="StyleBoxFlat" id="StyleBoxFlat_e20w6"]
-
-[node name="MainMenu" type="Node2D" node_paths=PackedStringArray("LoadDialog", "SetCiv3Home", "SetCiv3HomeDialog", "LoadScenarioDialog")]
-z_index = 1
-z_as_relative = false
-script = ExtResource("1")
-LoadDialog = NodePath("CanvasLayer/LoadDialog")
-SetCiv3Home = NodePath("CanvasLayer/SetCiv3Home")
-SetCiv3HomeDialog = NodePath("CanvasLayer/SetCiv3HomeDialog")
-LoadScenarioDialog = NodePath("CanvasLayer/LoadScenarioDialog")
-
-[node name="CanvasLayer" type="CanvasLayer" parent="."]
-layer = -1
-
-[node name="ColorRect" type="ColorRect" parent="CanvasLayer"]
-visible = false
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-grow_horizontal = 2
-grow_vertical = 2
-size_flags_horizontal = 3
-size_flags_vertical = 3
-color = Color(0.286275, 0.592157, 0.580392, 1)
-
-[node name="Label" type="Label" parent="CanvasLayer"]
-visible = false
-anchors_preset = 8
-anchor_left = 0.5
-anchor_top = 0.5
-anchor_right = 0.5
-anchor_bottom = 0.5
-offset_left = -160.0
-offset_top = -54.0
-offset_right = 160.0
-offset_bottom = 54.0
-grow_horizontal = 2
-grow_vertical = 2
-text = "Civfanatics Creation &
-Customization Community's
-Civ3 Conquests Clone
-Unable to load default graphics
-On Windows, these should be auto-detected from
-the Registry if Civ3 is installed.
-On Mac and Linux, set the environment variable
-CIV3_HOME to the Civ install location.
-Then restart C7."
-
-[node name="MainMenuMusicPlayer" type="AudioStreamPlayer" parent="CanvasLayer"]
-script = ExtResource("3")
-
-[node name="SoundEffectPlayer" type="AudioStreamPlayer" parent="CanvasLayer"]
-
-[node name="MainMenuBackground" type="TextureRect" parent="CanvasLayer"]
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-grow_horizontal = 2
-grow_vertical = 2
-theme = ExtResource("4")
-stretch_mode = 1
-script = ExtResource("5")
-
-[node name="SetCiv3Home" type="Button" parent="CanvasLayer"]
-anchors_preset = 8
-anchor_left = 0.5
-anchor_top = 0.5
-anchor_right = 0.5
-anchor_bottom = 0.5
-offset_left = -155.0
-offset_top = -101.0
-offset_right = 159.0
-offset_bottom = -35.0
-grow_horizontal = 2
-grow_vertical = 2
-theme_override_constants/h_separation = 50
-theme_override_font_sizes/font_size = 24
-theme_override_styles/normal = SubResource("StyleBoxFlat_gmerc")
-theme_override_styles/hover = SubResource("StyleBoxFlat_vbij8")
-theme_override_styles/pressed = SubResource("StyleBoxFlat_e20w6")
-text = "Select Civ3 Home Folder"
-
-[node name="SetCiv3HomeDialog" type="FileDialog" parent="CanvasLayer"]
-mode = 3
-title = "Open a Directory"
-initial_position = 1
-size = Vector2i(357, 450)
-ok_button_text = "Select Current Folder"
-file_mode = 2
-access = 2
-
-[node name="LoadDialog" parent="CanvasLayer" instance=ExtResource("4_x23yr")]
-visible = false
-
-[node name="LoadScenarioDialog" parent="CanvasLayer" instance=ExtResource("4_x23yr")]
-visible = false
-
-[connection signal="pressed" from="CanvasLayer/SetCiv3Home" to="." method="_on_SetCiv3Home_pressed"]
-[connection signal="dir_selected" from="CanvasLayer/SetCiv3HomeDialog" to="." method="_on_SetCiv3HomeDialog_dir_selected"]
