+++ C7/UIElements/MainMenu/MainMenuTheme.tres
@@ -1,0 +1,32 @@
+[gd_resource type="Theme" load_steps=6 format=3 uid="uid://b0jisy3avwkxf"]
+
+[sub_resource type="StyleBoxEmpty" id="3"]
+
+[sub_resource type="StyleBoxEmpty" id="4"]
+
+[sub_resource type="StyleBoxEmpty" id="2"]
+
+[sub_resource type="StyleBoxEmpty" id="5"]
+
+[sub_resource type="StyleBoxFlat" id="1"]
+bg_color = Color(1, 0.968627, 0.870588, 1)
+expand_margin_left = 2.0
+expand_margin_top = 3.0
+expand_margin_right = 2.0
+expand_margin_bottom = 3.0
+
+[resource]
+Button/colors/font_color = Color(0.1, 0.1, 0.1, 1)
+Button/colors/font_color_focus = Color(0, 0.45098, 0.741176, 1)
+Button/colors/font_color_hover = Color(0.709804, 0.352941, 0, 1)
+Button/colors/font_color_pressed = Color(0, 0.45098, 0.741176, 1)
+Button/styles/focus = SubResource("3")
+Button/styles/hover = SubResource("4")
+Button/styles/normal = SubResource("2")
+Button/styles/pressed = SubResource("5")
+Label/colors/font_color = Color(0, 0, 0, 1)
+LineEdit/colors/cursor_color = Color(0, 0, 0, 1)
+LineEdit/colors/font_color = Color(0, 0, 0, 1)
+LineEdit/colors/selection_color = Color(0.709804, 0.807843, 0.709804, 1)
+LineEdit/styles/focus = SubResource("1")
+LineEdit/styles/normal = SubResource("1")
