+++ C7/UIElements/MainMenu/main_menu.tscn
@@ -1,0 +1,105 @@
+[gd_scene load_steps=9 format=3 uid="uid://gnt0y3og7nk6"]
+
+[ext_resource type="Script" path="res://UIElements/MainMenu/MainMenu.cs" id="1_akneh"]
+[ext_resource type="Texture2D" uid="uid://bu113i1ououof" path="res://Art/title-screen.png" id="2_1lqrk"]
+[ext_resource type="Script" path="res://UIElements/MainMenu/MenuButtonContainer.cs" id="3_t1nuq"]
+[ext_resource type="Script" path="res://UIElements/MainMenu/MainMenuMusicPlayer.cs" id="4_tqr50"]
+[ext_resource type="PackedScene" uid="uid://bvncm1bgcii5e" path="res://UIElements/civ3_file_dialog.tscn" id="5_rw2g7"]
+
+[sub_resource type="StyleBoxFlat" id="StyleBoxFlat_gmerc"]
+
+[sub_resource type="StyleBoxFlat" id="StyleBoxFlat_vbij8"]
+
+[sub_resource type="StyleBoxFlat" id="StyleBoxFlat_e20w6"]
+
+[node name="MainMenu" type="Node" node_paths=PackedStringArray("LoadDialog", "SetCiv3Home", "SetCiv3HomeDialog", "LoadScenarioDialog", "ButtonContainer", "player")]
+script = ExtResource("1_akneh")
+LoadDialog = NodePath("LoadDialog")
+SetCiv3Home = NodePath("SetCiv3Home")
+SetCiv3HomeDialog = NodePath("SetCiv3HomeDialog")
+LoadScenarioDialog = NodePath("LoadScenarioDialog")
+ButtonContainer = NodePath("Control/MainMenuContainer")
+player = NodePath("SoundEffectPlayer")
+
+[node name="Control" type="Control" parent="."]
+layout_mode = 3
+anchors_preset = 15
+anchor_right = 1.0
+anchor_bottom = 1.0
+grow_horizontal = 2
+grow_vertical = 2
+size_flags_vertical = 3
+
+[node name="Background" type="TextureRect" parent="Control"]
+layout_mode = 1
+anchors_preset = 13
+anchor_left = 0.5
+anchor_right = 0.5
+anchor_bottom = 1.0
+offset_left = -419.669
+offset_right = 419.669
+grow_horizontal = 2
+grow_vertical = 2
+size_flags_horizontal = 3
+size_flags_vertical = 3
+texture = ExtResource("2_1lqrk")
+expand_mode = 3
+stretch_mode = 4
+
+[node name="MainMenuContainer" type="VBoxContainer" parent="Control"]
+layout_mode = 1
+anchors_preset = -1
+anchor_left = 0.3
+anchor_top = 0.1
+anchor_right = 0.5
+anchor_bottom = 0.2
+offset_left = -79.0
+offset_right = 79.0
+offset_bottom = 370.0
+grow_horizontal = 2
+theme_override_constants/separation = 15
+script = ExtResource("3_t1nuq")
+
+[node name="MainMenuMusicPlayer" type="AudioStreamPlayer" parent="."]
+script = ExtResource("4_tqr50")
+
+[node name="SoundEffectPlayer" type="AudioStreamPlayer" parent="."]
+
+[node name="SetCiv3Home" type="Button" parent="."]
+anchors_preset = 8
+anchor_left = 0.5
+anchor_top = 0.5
+anchor_right = 0.5
+anchor_bottom = 0.5
+offset_left = -157.0
+offset_top = -33.0
+offset_right = 157.0
+offset_bottom = 33.0
+grow_horizontal = 2
+grow_vertical = 2
+size_flags_horizontal = 4
+size_flags_vertical = 4
+theme_override_constants/h_separation = 50
+theme_override_font_sizes/font_size = 24
+theme_override_styles/normal = SubResource("StyleBoxFlat_gmerc")
+theme_override_styles/hover = SubResource("StyleBoxFlat_vbij8")
+theme_override_styles/pressed = SubResource("StyleBoxFlat_e20w6")
+text = "Select Civ3 Home Folder"
+
+[node name="SetCiv3HomeDialog" type="FileDialog" parent="."]
+mode = 3
+title = "Open a Directory"
+initial_position = 1
+size = Vector2i(357, 450)
+ok_button_text = "Select Current Folder"
+file_mode = 2
+access = 2
+
+[node name="LoadDialog" parent="." instance=ExtResource("5_rw2g7")]
+visible = false
+
+[node name="LoadScenarioDialog" parent="." instance=ExtResource("5_rw2g7")]
+visible = false
+
+[connection signal="pressed" from="SetCiv3Home" to="." method="_on_SetCiv3Home_pressed"]
+[connection signal="dir_selected" from="SetCiv3HomeDialog" to="." method="_on_SetCiv3HomeDialog_dir_selected"]
